Fix filter queryparam not being removed after clearing the filter
It's a bit of a guess because I can't replicate the issue locally, but isEnabled sends 'undefined' when filter is disabled and it works fine. So I assume that sending 'undefined' should also help in onChange.
Also got rid of 'temp fix' in onChange - hiding the drop-down menu makes text disappear so it wasn't required anymore